@@ -3,7 +3,7 @@ description: Demonstrates how to use the cannon_physics plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.17.6 <4.0.0"
+  sdk: ">=3.3.3 <4.0.0"
 
 dependencies:
   flutter:
@@ -12,13 +12,8 @@ dependencies:
   cannon_physics:
     path: ../
   vector_math: ^2.1.4
-  flutter_gl: ^0.0.21
-  three_dart:
-    git:
-      url: https://github.com/Knightro63/three_dart.git
-  three_dart_jsm:
-    git:
-      url: https://github.com/Knightro63/three_dart_jsm.git
+  three_js: ^0.0.5
+  three_js_geometry: ^0.0.2
   css:
     git:
       url: https://github.com/Knightro63/css.git
