@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../src/demo.dart';
-import 'package:three_dart/three_dart.dart';
+import 'dart:math' as math;
 import 'package:cannon_physics/cannon_physics.dart' as cannon;
 import 'package:vector_math/vector_math.dart' as vmath;
 
@@ -47,7 +47,7 @@ class _BodyTypesState extends State<BodyTypes> {
     final groundShape = cannon.Plane();
     final groundBody = cannon.Body(mass: 0 );
     groundBody.addShape(groundShape);
-    groundBody.quaternion.setFromEuler(-Math.PI / 2, 0, 0);
+    groundBody.quaternion.setFromEuler(-math.pi / 2, 0, 0);
     world.addBody(groundBody);
     demo.addVisual(groundBody);
 
